refactore: Organise route imports and registration
