+++ src/routes/index.ts
@@ -1,4 +1,6 @@
 import home from './home';
 import ping from './ping';
+import createSession from './createSession';
 
-export default [home, ping];
+export const get = [home, ping];
+export const post = [createSession];
